refactor(frontend): comprehensive architecture refactoring
## Summary
Major frontend code refactoring for improved code organization,
maintainability, and adherence to best practices.

## Key Changes

### Component Decomposition
- Split main-sidebar.tsx (828→180 lines) into:
  sidebar-header.tsx, sidebar-footer.tsx, sidebar-content.tsx,
  hooks/use-sidebar-selection.ts
- Split task-composer.tsx (878→350 lines) into:
  composer-toolbar.tsx, composer-attachments.tsx, repo-dialog.tsx,
  slash-autocomplete-dropdown.tsx, hooks/use-file-upload.ts

### Service Layer Refactoring
- Extracted message-parser.ts (357 lines) from chat-service.ts
- Extracted file-tree-builder.ts (139 lines) from chat-service.ts
- chat-service.ts reduced from 610→250 lines (pure API orchestration)

### Infrastructure Improvements
- Renamed proxy.ts (Next.js 16 convention, middleware→proxy)
- Renamed CardNav.tsx → card-nav.tsx (kebab-case convention)
- Renamed features/home/model/ → features/home/constants/
- Consolidated storage utilities (2 files → 1 type-safe module)
- Removed duplicate use-mobile hook
- Cleaned up api-client.ts (removed dead code, improved logging)

### Style & Type System
- Reorganized globals.css with clear section comments
- Removed deprecated skeleton animations
- Added JSDoc comments to shared type definitions
- Improved types/index.ts barrel exports

## Verification
- TypeScript compilation: 0 errors
- Next.js production build: successful (all 28 pages generated)
