+++ frontend/features/chat/components/chat/model-selector.tsx
@@ -9,7 +9,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { Button } from "@/components/ui/button";
-import { AVAILABLE_MODELS } from "@/features/home/model/constants";
+import { AVAILABLE_MODELS } from "@/features/home/constants/constants";
 import type { ModelInfo } from "@/types";
 import { useT } from "@/lib/i18n/client";
