+++ frontend/lib/api-client.ts
@@ -1,8 +1,18 @@
 import { ApiError } from "./errors";
 import type { ApiResponse } from "@/types";
 
+// ---------------------------------------------------------------------------
+// Constants
+// ---------------------------------------------------------------------------
+
 export const API_PREFIX = "/api/v1";
 
+/**
+ * Centralized API endpoint definitions.
+ *
+ * Static endpoints are plain strings; dynamic endpoints are factory functions
+ * that accept the required identifier and return the path segment.
+ */
 export const API_ENDPOINTS = {
   // Sessions
   sessions: "/sessions",
@@ -23,7 +33,7 @@ export const API_ENDPOINTS = {
   sessionWorkspaceArchive: (sessionId: string) =>
     `/sessions/${sessionId}/workspace/archive`,
 
-  // User Input Requests (AskUserQuestion)
+  // User Input Requests
   userInputRequests: "/user-input-requests",
   userInputRequest: (requestId: string) => `/user-input-requests/${requestId}`,
   userInputAnswer: (requestId: string) =>
@@ -35,7 +45,7 @@ export const API_ENDPOINTS = {
   // Runs
   runsBySession: (sessionId: string) => `/runs/session/${sessionId}`,
 
-  // User-level persistent custom instructions
+  // Custom Instructions
   customInstructions: "/claude-md",
 
   // Attachments
@@ -49,7 +59,7 @@ export const API_ENDPOINTS = {
   mcpServers: "/mcp-servers",
   mcpServer: (serverId: number) => `/mcp-servers/${serverId}`,
 
-  // MCP Installs (User)
+  // MCP Installs
   mcpInstalls: "/mcp-installs",
   mcpInstall: (installId: number) => `/mcp-installs/${installId}`,
   mcpInstallsBulk: "/mcp-installs/bulk",
@@ -61,7 +71,7 @@ export const API_ENDPOINTS = {
   skillImportCommit: "/skills/import/commit",
   skillImportJob: (jobId: string) => `/skills/import/jobs/${jobId}`,
 
-  // Skill Installs (User)
+  // Skill Installs
   skillInstalls: "/skill-installs",
   skillInstall: (installId: number) => `/skill-installs/${installId}`,
   skillInstallsBulk: "/skill-installs/bulk",
@@ -73,7 +83,7 @@ export const API_ENDPOINTS = {
   pluginImportCommit: "/plugins/import/commit",
   pluginImportJob: (jobId: string) => `/plugins/import/jobs/${jobId}`,
 
-  // Plugin Installs (User)
+  // Plugin Installs
   pluginInstalls: "/plugin-installs",
   pluginInstall: (installId: number) => `/plugin-installs/${installId}`,
   pluginInstallsBulk: "/plugin-installs/bulk",
@@ -103,41 +113,56 @@ export const API_ENDPOINTS = {
   // Tool Executions
   toolExecution: (executionId: string) => `/tool-executions/${executionId}`,
 
-  // Other
+  // Projects
   projects: "/projects",
   project: (projectId: string) => `/projects/${projectId}`,
+
+  // Health
   health: "/health",
   root: "/",
-};
+} as const;
 
-export const API_BASE_URL = process.env.NEXT_PUBLIC_API_URL;
+// ---------------------------------------------------------------------------
+// Base URL resolution
+// ---------------------------------------------------------------------------
+
+const API_BASE_URL = process.env.NEXT_PUBLIC_API_URL;
 
 function normalizeBaseUrl(baseUrl: string): string {
   return baseUrl.endsWith("/") ? baseUrl.slice(0, -1) : baseUrl;
 }
 
-export function getApiBaseUrl() {
-  // Browser: prefer an explicit public API URL (build-time), otherwise use same-origin proxy.
+/**
+ * Resolve the API base URL for the current environment.
+ *
+ * - **Browser**: uses `NEXT_PUBLIC_API_URL` or falls back to same-origin proxy.
+ * - **Server**: requires `BACKEND_URL` / `POCO_BACKEND_URL` for absolute URLs.
+ */
+export function getApiBaseUrl(): string {
   if (typeof window !== "undefined") {
-    if (!API_BASE_URL) return "";
-    return normalizeBaseUrl(API_BASE_URL);
+    return API_BASE_URL ? normalizeBaseUrl(API_BASE_URL) : "";
   }
 
-  // Server: relative URLs are not supported by Node.js fetch. Prefer a runtime internal URL.
   const serverBaseUrl =
     process.env.BACKEND_URL || process.env.POCO_BACKEND_URL || API_BASE_URL;
+
   if (!serverBaseUrl) {
     throw new ApiError(
       "API base URL is not configured (set BACKEND_URL for server-side calls)",
       500,
     );
   }
+
   return normalizeBaseUrl(serverBaseUrl);
 }
 
+// ---------------------------------------------------------------------------
+// Auth token resolution
+// ---------------------------------------------------------------------------
+
 async function resolveAuthToken(): Promise<string | null> {
-  // TODO: add token support
   if (typeof window === "undefined") {
+    // Server-side: read from cookies via next/headers
     try {
       const { cookies } = await import("next/headers");
       const cookieStore = await cookies();
@@ -151,10 +176,7 @@ async function resolveAuthToken(): Promise<string | null> {
     }
   }
 
-  if (typeof window === "undefined") {
-    return null;
-  }
-
+  // Client-side: read from localStorage
   try {
     return (
       window.localStorage.getItem("access_token") ||
@@ -166,11 +188,38 @@ async function resolveAuthToken(): Promise<string | null> {
   }
 }
 
+// ---------------------------------------------------------------------------
+// Request helpers
+// ---------------------------------------------------------------------------
+
 function normalizeBody(body: unknown): BodyInit | undefined {
   if (body === undefined || body === null) return undefined;
   if (typeof body === "string" || body instanceof FormData) return body;
   return JSON.stringify(body);
 }
+
+const IS_DEV =
+  typeof process !== "undefined" && process.env.NODE_ENV === "development";
+
+function logApi(
+  level: "info" | "error",
+  method: string,
+  endpoint: string,
+  extra?: Record<string, unknown>,
+): void {
+  if (!IS_DEV) return;
+  const color = level === "error" ? "#ef4444" : "#0ea5e9";
+  const prefix = `%c[API] ${method} ${endpoint}`;
+  if (extra) {
+    console.log(prefix, `color: ${color}; font-weight: bold;`, extra);
+  } else {
+    console.log(prefix, `color: ${color}; font-weight: bold;`);
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Core fetch wrapper
+// ---------------------------------------------------------------------------
 
 export type ApiFetchOptions = RequestInit & {
   timeoutMs?: number;
@@ -180,6 +229,14 @@ export type ApiFetchOptions = RequestInit & {
   };
 };
 
+/**
+ * Low-level fetch wrapper that handles:
+ * - Base URL resolution and endpoint prefixing
+ * - Auth token injection
+ * - Timeout via AbortController
+ * - Unwrapping `ApiResponse<T>` envelopes
+ * - Structured error handling
+ */
 export async function apiFetch<T>(
   endpoint: string,
   options: ApiFetchOptions = {},
@@ -187,18 +244,13 @@ export async function apiFetch<T>(
   const { timeoutMs = 60_000, ...fetchOptions } = options;
   const baseUrl = getApiBaseUrl();
   const fullUrl = `${baseUrl}${API_PREFIX}${endpoint}`;
+  const method = fetchOptions.method || "GET";
 
-  console.log(
-    `%c[API] ${options.method || "GET"} ${endpoint}`,
-    "color: #0ea5e9; font-weight: bold;",
-  );
+  logApi("info", method, endpoint);
 
+  // Build headers
   const headers = new Headers(fetchOptions.headers);
-
-  if (
-    !headers.has("Content-Type") &&
-    !(fetchOptions.body instanceof FormData)
-  ) {
+  if (!headers.has("Content-Type") && !(fetchOptions.body instanceof FormData)) {
     headers.set("Content-Type", "application/json");
   }
 
@@ -207,6 +259,7 @@ export async function apiFetch<T>(
     headers.set("Authorization", `Bearer ${token}`);
   }
 
+  // Timeout management
   const hasExternalSignal = Boolean(fetchOptions.signal);
   const controller = hasExternalSignal ? null : new AbortController();
   const timeoutId =
@@ -232,11 +285,10 @@ export async function apiFetch<T>(
           ? String((payload as { message?: string }).message)
           : response.statusText;
 
-      console.log(
-        `%c[API] Error ${endpoint}`,
-        "color: #ef4444; font-weight: bold;",
-        { status: response.status, message, payload },
-      );
+      logApi("error", method, endpoint, {
+        status: response.status,
+        message,
+      });
 
       throw new ApiError(
         message || "API request failed",
@@ -245,12 +297,7 @@ export async function apiFetch<T>(
       );
     }
 
-    console.log(
-      `%c[API] Success ${endpoint}`,
-      "color: #22c55e; font-weight: bold;",
-      payload,
-    );
-
+    // Unwrap standard API envelope { code, message, data }
     if (typeof payload === "object" && payload && "data" in payload) {
       const wrapped = payload as ApiResponse<T>;
       if (wrapped.code !== 200 && wrapped.code !== 0) {
@@ -264,42 +311,56 @@ export async function apiFetch<T>(
 
     return payload as T;
   } catch (error) {
-    const errorName =
-      typeof error === "object" && error && "name" in error
-        ? String((error as { name?: unknown }).name)
-        : "";
-    if (errorName === "AbortError") {
+    if (
+      error instanceof DOMException &&
+      error.name === "AbortError"
+    ) {
       throw new ApiError("Request timeout", 408);
     }
     throw error;
   } finally {
-    if (timeoutId) {
-      clearTimeout(timeoutId);
-    }
+    if (timeoutId) clearTimeout(timeoutId);
   }
 }
 
+// ---------------------------------------------------------------------------
+// Typed HTTP client
+// ---------------------------------------------------------------------------
+
+/**
+ * Convenience HTTP client with typed methods for each verb.
+ *
+ * @example
+ * ```ts
+ * const sessions = await apiClient.get<Session[]>(API_ENDPOINTS.sessions);
+ * await apiClient.post(API_ENDPOINTS.sessions, { prompt: "Hello" });
+ * ```
+ */
 export const apiClient = {
   get: <T>(endpoint: string, options?: ApiFetchOptions) =>
     apiFetch<T>(endpoint, { ...options, method: "GET" }),
+
   post: <T>(endpoint: string, body?: unknown, options?: ApiFetchOptions) =>
     apiFetch<T>(endpoint, {
       ...options,
       method: "POST",
       body: body as BodyInit | null | undefined,
     }),
+
   patch: <T>(endpoint: string, body?: unknown, options?: ApiFetchOptions) =>
     apiFetch<T>(endpoint, {
       ...options,
       method: "PATCH",
       body: body as BodyInit | null | undefined,
     }),
+
   put: <T>(endpoint: string, body?: unknown, options?: ApiFetchOptions) =>
     apiFetch<T>(endpoint, {
       ...options,
       method: "PUT",
       body: body as BodyInit | null | undefined,
     }),
+
   delete: <T>(endpoint: string, options?: ApiFetchOptions) =>
     apiFetch<T>(endpoint, { ...options, method: "DELETE" }),
 };
