--- frontend/lib/hooks/use-mobile.ts
@@ -1,1 +1,0 @@
-export { useIsMobile } from "@/hooks/use-mobile";
