+++ frontend/types/index.ts
@@ -1,3 +1,10 @@
+/**
+ * Global type definitions shared across the entire application.
+ *
+ * Feature-specific types should be imported from their respective
+ * feature module (e.g., `@/features/chat/types`).
+ */
+
 export * from "./api";
 export * from "./pagination";
 export * from "./ui";
